% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BioC2015OpenCyto.R
\docType{package}
\name{BioC2015OpenCyto}
\alias{BioC2015OpenCyto}
\alias{BioC2015OpenCyto-package}
\title{BioC2015OpenCyto}
\description{
Data and code demonstrating the analysis of designed experiments using OpenCyto.
}
\details{
BioC2015OpenCyto

Analyzing Designed Experiments with OpenCyto

Demonstrates how to use  the OpenCyto framework to gate FCM data utilizing study metadata to structure the gating to use controls
}
\seealso{
\link{tbdata}
}

